fix: resolve remaining Dart compile errors breaking the Android build
- Wrap the Fandoghi widget in a SizedBox before calling .animate(): the
  Fandoghi widget exposes an 'animate' bool field which shadows the
  flutter_animate .animate() extension, so calling it directly on the
  widget was 'not a function/method'.
- Revert the x:.toDouble() on BarChartGroupData.x (main.dart and
  profile_screen.dart): fl_chart 0.69.2 expects x to be an int, so passing
  a double was a compile error.

+++ lib/screens/profile_screen.dart
@@ -414,7 +414,7 @@ class _ProfileScreenState extends State<ProfileScreen> {
                       .asMap()
                       .entries
                       .map((e) => BarChartGroupData(
-                            x: e.key.toDouble(),
+                            x: e.key,
                             barRods: [
                               BarChartRodData(
                                 toY: e.value.value.toDouble(),
